LOADER: move more code, simplify `ModuleCallback`
