+++ base/src/main/java/org/aya/concrete/resolve/module/ModuleLoader.java
@@ -3,12 +3,15 @@
 package org.aya.concrete.resolve.module;
 
 import kala.collection.immutable.ImmutableSeq;
+import org.aya.api.error.DelayedReporter;
 import org.aya.api.error.Reporter;
 import org.aya.concrete.desugar.AyaBinOpSet;
 import org.aya.concrete.resolve.ModuleCallback;
 import org.aya.concrete.resolve.ResolveInfo;
 import org.aya.concrete.resolve.context.ModuleContext;
 import org.aya.concrete.stmt.Stmt;
+import org.aya.tyck.order.AyaNonStoppingTicker;
+import org.aya.tyck.order.AyaSccTycker;
 import org.aya.tyck.trace.Trace;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -24,11 +27,21 @@ public interface ModuleLoader {
     @Nullable ModuleCallback<E> onTycked
   ) throws E {
     var resolveInfo = resolveModule(context, program);
-    FileModuleLoader.tyckResolvedModule(resolveInfo, reporter(), builder, onTycked);
+    var delayedReporter = new DelayedReporter(reporter());
+    var sccTycker = new AyaNonStoppingTicker(new AyaSccTycker(builder, delayedReporter), resolveInfo);
+    // in case we have un-messaged TyckException
+    try (delayedReporter) {
+      var SCCs = resolveInfo.declGraph().topologicalOrder()
+        .view().appendedAll(resolveInfo.sampleGraph().topologicalOrder())
+        .toImmutableSeq();
+      SCCs.forEach(sccTycker::tyckSCC);
+    } finally {
+      if (onTycked != null) onTycked.onModuleTycked(
+        resolveInfo, sccTycker.sccTycker().wellTyped().toImmutableSeq());
+    }
     return resolveInfo;
   }
 
-  @NotNull Reporter reporter();
   default @NotNull ResolveInfo resolveModule(
     @NotNull ModuleContext context,
     @NotNull ImmutableSeq<Stmt> program
@@ -38,5 +51,6 @@ public interface ModuleLoader {
     return resolveInfo;
   }
 
+  @NotNull Reporter reporter();
   @Nullable ResolveInfo load(@NotNull ImmutableSeq<@NotNull String> path);
 }
