+++ base/src/main/java/org/aya/concrete/resolve/visitor/StmtResolver.java
@@ -81,7 +81,7 @@ public interface StmtResolver {
         resolveDeclSignature(decl, ExprResolver.RESTRICTIVE)._1);
       case Sample sample -> {
         var delegate = sample.delegate();
-        var delegateInfo = new ResolveInfo(info.thisModule(), info.thisProgram(), info.opSet());
+        var delegateInfo = new ResolveInfo(info.thisModule(), info.program(), info.opSet());
         resolveStmt(delegate, delegateInfo);
         // little hacky: transfer dependencies from `delegate` to `sample`
         info.sampleGraph().suc(sample).appendAll(delegateInfo.declGraph().suc(delegate));
