+++ cli/src/main/java/org/aya/cli/repl/ReplCompiler.java
@@ -63,7 +63,7 @@ public class ReplCompiler {
       return programOrExpr.map(
         program -> {
           var newDefs = new Ref<ImmutableSeq<Def>>();
-          loader.tyckModule(context, program, null, ((moduleResolve, stmts, defs) -> newDefs.set(defs)));
+          loader.tyckModule(context, program, null, ((moduleResolve, defs) -> newDefs.set(defs)));
           var defs = newDefs.get();
           if (reporter.noError()) return defs;
           else {
